refactor: Add only valid `orderBy` clauses to `findAllPlayers`
The orderBy clause now uses a `reduce` instead of a `map`. This allows
us to check if a field really exists i.e. has a valid orderBy direction,
instead of sorting it by default.
This was a potential issue earlier, since our `orderBy` field is of type
`Partial<Record<TColumnName, ESortDirection>>`