fix: Order `findAllPlayers` based on query, not network responses
Earlier, we used a `.push` inside an async map that added to the list of
values one by one as they came in after being issued in parallel. This
screws up the order provided by the initial `players` table query.

So instead now we fetch the counts data, store in a hash table, and then
simply augment the initial data while preserving the order.