chore: Remove TODO comments re. entries-entrants type & transformation hooks
After research, it was found that there is no native pre-insert
transform hook in drizzle. This is to be handled on the application
layer when we create the service.

Also, solana addresses do not have an `Address` type as in viem. The
base58 encoded strings are of type `string`. So no further ornamentation
is required