86c093jxw | Setup CI/CD deployment and tests (#22)
* ci: Add deploy pipeline

The pipeline temporarily set to deploy on push to current branch to enable
testing. This will be updated to `on_workflow_dispatch` in the future.

Also the make commands were updated to use the exact CDK version from
package.json.

* misc: Add environment and .env file to deploy-prod workflow

* ci: Update .env variables in pipeline and Makefile

The AWS_PROFILE was removed in favour of using the default profile
configured in CI/CD pipeline.

* ci: Update deploy pipeline to run only on main

* ci: Add test and typechecking workflow

It runs on every PR

* chore: Fix tsc build errors

The lib (node_module) typechecking was disabled as it is outside our
application's ambit. Any other outstanding type errors were fixed.

* test: Fix getPlayersHandler test

The incorrect value earlier expected from the logger.error, was fixed

* ci: Ignore `saveTwitterProfile` test suite

This test suite has some issues running without env variables. So we
ignore it in the testing pipelines until we can modify the tests to work
without env set up.