test&ci: Fix twitter tests by mocking `getMandatoryEnvValue`
As noted in the code comments, this is a bad practice in any non-service
layer tests. In higher level module tests like this, we should mock the
whole service instead.

However it's a quick and dirty fix, and less intense than debugging the
whole twitter login flow (problematic due to `twitter-api-sdk`'s `authClient`
being stateful).

We also un-skipped this test on the CI test job