feat: Update PlayersService in preparation for twitter-connect points
When the user connects their twitter account, we need to award them points.
In anticipation of this, we need to modify our existing PlayersService methods
so that we first check if the user exists, and based on that either add a
new player or update the existing one's twitter info.

For this purpose, the `insertNew` and `changePoints` methods were added,
and the old `insertNewOrSaveTwitterProfile` was modified as `updateTwitterProfile`.
As part of these changes, the type signature on `getPlayerById` was
explicitly set, and the `updateTwitterProfile` was modified such that it
does _not_ return any value to the caller (command query separation).