86c0b0h3m | Award points for connecting twitter  (#28)
* feat: Update PlayersService in preparation for twitter-connect points

When the user connects their twitter account, we need to award them points.
In anticipation of this, we need to modify our existing PlayersService methods
so that we first check if the user exists, and based on that either add a
new player or update the existing one's twitter info.

For this purpose, the `insertNew` and `changePoints` methods were added,
and the old `insertNewOrSaveTwitterProfile` was modified as `updateTwitterProfile`.
As part of these changes, the type signature on `getPlayerById` was
explicitly set, and the `updateTwitterProfile` was modified such that it
does _not_ return any value to the caller (command query separation).

* feat: Add points for linking twitter

- When the user hits '/save-twitter-profile' clientApi endpoint, we check
  if they are in db. If they aren't we add them (with twitter profile data)
  and award them 10 points. If they are in the db, they still get 10 points
  if there's no twitter profile yet. But if they've already linked the
  twitter profile we just update the profile.

- We moved the `findHighResImageUrl` to a shared util. This will aid in
  testing later.

- Minor control flow and logging improvements were also made.

* test: Modify `saveTwitterProfile` handler tests with points award logic

The earlier testcases were split up into smaller units, to handle each
control flow case. Some logic such as the `verifySignature` were tested in
it's respective _cryptography.ts_ module.

* chore: Rename `getPlayerById` to `getPlayerByAddress` in `PlayersService`

The earlier name was getting confusing as we also have a `getPlayerByTwitterId`
function. Plus, there is no real field called `id` in the `players` table.
The new method name makes it clear that it's fetching by address.

* chore: Refactor `saveTwitterProfile` logic and tests

The order of variable declarations was slightly adjusted in the actual
module (to create variables just before using).

In the associated tests, `mockEvent` was abstracted and shared between tests.

* test: Add tests for `saveTwitterProfile` when user has/hasn't linked twitter yet

Earlier, we only had one test for when the user was in the db. Now we
split that into two - for when the user has and hasn't linked twitter yet.

As an aside jest.clearAllMocks() was added in the beforeEach() of `poolEntered`
test

* fix: Update status codes to 400 for bad requests in `saveTwitterProfile`

Earlier we used the internal server error code by mistake.