feat: Add `verifyTwitterShare` handler
The handler verifies if a given tweet contains a link to a pool page.
If it does, it means the user has successfully shared the pool, so we
give them points. Else we don't. We return a boolean indicating if the
verification of the share was successful.

Storing of the verified shares in a different table is not yet implemented