86c0b0pr8 | `/claim-pool-tweet-points` endpoint  (#35)
* feat: Add `findTweetContentById` to Twitter utils

This helps us find the content of a tweet by fetching it with its tweet ID.

* feat: Add `parseTweetIdFromUrl` and `getPoolPageUrlFromPoolId` utils

These will be later used in `verifyTwitterShare`. Tests were also added
for these utils.

* feat: Add `verifyTwitterShare` handler

The handler verifies if a given tweet contains a link to a pool page.
If it does, it means the user has successfully shared the pool, so we
give them points. Else we don't. We return a boolean indicating if the
verification of the share was successful.

Storing of the verified shares in a different table is not yet implemented

* chore: Move `playersTable` types into `players/schema.ts`

No point having them in a different file. There's no cyclic dependency

* feat: Add poolSharingTweetsTable and generate migrations

We will store the tweets containing pool url shared by users via
`verifyTwitterShareHandler` in this table. This helps us avoid duplicate
verifications of the same tweet, thus preventing point farming

* feat: Update `verifyTwitterShareHandler` to insert into `poolSharingTweets` table

If the tweet is legit, we insert it into the `poolSharingTweets` table.
The service method `insertNew` was added for this purpose.

Tests were also updated to satisfy the new functionality. Minor refactors
were also done in the test file

* chore: Rename fields on poolSharingTweets table

The new names are less verbose, and get the job done. The index names
were also updated.

* feat: Add check for duplicate tweet shares in `verifyTwitterShareHandler`

If a request tries to share a tweet that has already been used to claim
points, the handler will now return a bad request error. We need this
to avoid people from farming points by sharing the same tweet multiple times.

* test: Update `verifyTwitterShareHandler` tests

An `it` block was added to test the case where a tweet has already been
verified. Also the checks on `insertNew` method of `PoolSharingTweetsService`
were updated to use the correct argument names.

* feat: Rename `verifyTwitterShare` to `claimPoolTweetPoints`

This is more descriptive. The HTTP endpoint was also renamed, as were the
handler and test file names

* feat: Add logging to `claimPoolTweetPoints`

* refactor: Modify `claimPoolTweetPoints` to return 400 if tweet doesn't contain pool URL

If the claim failed because the tweet doesn't contain the pool URL,
we should return 400. This means we have departed from the earlier standard
of returning 200 with `{isSuccess: boolean}` body.

The testcases were also updated to reflect the new expected behavior.

* feat: Update `claimPoolTweetPoints` to return the points awarded also

If the claim was successful, we also return the number of points awarded
to the player, in the response body.

* test: Mock logger centrally in test setup file

* fix: Address review comments

https://github.com/Degen-Markets/degen-markets-be/pull/35#discussion_r1768285416
https://github.com/Degen-Markets/degen-markets-be/pull/35#discussion_r1768328937
https://github.com/Degen-Markets/degen-markets-be/pull/35#discussion_r1768339066

* test: Add more tests for parseTweetIdFromUrl

We're checking the aspects related to \b in the regex

* fix: Use bearer token for `findTweetById` twitter util

The API requires bearer token, not the OAuth client. The env variable
was updated in all of the deployment flow as well