86c0hhp3p | Handle `WinnerSet` event  (#47)
* test: Replace `toBeCalled` with `toHaveBeenCalled`

The former is apparently deprecated.

* chore: Update IDL with `WinnerSet` event

We copied over the latest IDLs from `dgm-solana` repo

* fix: Strongly type decoded smart contract event in `SmartContractEventProcessor`

Earlier we had a lot of `any`'s in the types. With these changes, we
typecast the decoded event, which will prevent us from accessing invalid
fields on the event data

* feat: Handle `WinnerSet` event in `webhookApi` and `smartContractEventProcessor`

The events are received, transformed, and passed to the relevant event
handler.

* feat: Add logic to `winnerSet` event handler

When the event handler runs, we call `PoolOptionsService.setWinningOption`
to set the `isWinningOption` field on the specified option. This new
service method was added. Also, tests were added for the event handler.

* refactor: Add `SmartContractEventData` type

This abstracts the `Extract` logic, allowing us to easily access the
data type for a particular event. All references to the earlier type were
replaced with `SmartContractEventData`.

* fix: Remove unnecessary `VALID_EVENTS` check from `poolInteractionsHandler`

This check was unnecessary because we already run the events through
a switch case (in `mapLogToEventOrNull`) where we pass only the events we
care, and map the others to null.

* refactor: Move `decodeEventBase64Data` to utils file

This function gives us a typed version of anchor's `program.coder.events.decode`.
We moved it out of the types file and into a utils file, to keep the types
file just for types.

* test: Fix test with updated `PoolOptionsService.setWinner` call signature