86c0anthq | `/player:id/stats` endpoint  (#51)
* feat: Scaffold `players/:id/stats` endpoint

* feat: Add `getStats` method to `PlayersService`

This helps us source the stats for a player. Currently, this
is just the entries, with inner joins for pools and options

* feat: Add logic to `getPlayerStatsHandler`

We check if the received ID is for a legit player, then return
the stats for that player

* test: `getPlayerStatsHandler`

The handler was also made into a default export now

* test: Update `it` block to use `mockResolvedValueOnce`

Some `it` blocks were using `mockResolvedValue` instead
of the more temporal version. This could potentially cause
isues due to leaks into other test cases. So it was updated
codebase wide

* feat: Add option total value to `PlayersService.getStats`

This will help in calculating expected winnings

* fix: Update export for `getPlayerStatsHandler` in `clientApi`