feat: Upload pool images to S3
When user creates a pool with a custom image, we now save the image to
S3 and use that URL to create the pool on Solana.

For non-gif images, we convert it to png type, which ensures compatibility
with dialect, and also eliminates XSS vulnerabilities. GIF's we upload to
S3 as is.