fix: Replace `sharp` to fix lambda runtime issue
Sharp is a native library, so we are facing some mismatch between our
build/dev env (ubuntu) and prod (Amazon Linux 2 i.e. lambda) in terms of
dependencies.

In order to work around this, we now use the file-type package. The
`createPoolTx` handler logic was also modified to not try and convert
images to png, but rather use an allow-list of image types.

We are checking the image buffer to verify that it is indeed the specified
type instead of relying on MIME types like earlier.  Also we sanitize svgs.
These two measures help mitigate XSS attacks from S3 stored images.