86c0k8jda | Save pool images to S3  (#59)
* feat: Add bucket to `SolanaActionsStack`

As a start this will be used for storing users' pool images

* feat: Add S3 service

This abstracts away the complexity of dealing with S3 client library.
In order to provide the uploaded file URL to the consumer of this class,
we added a new env variable `AWS_REGION` to the `SolanaActionsStack`.

* feat: Revamp image service

The new service uses sharp, and has more powerful capabilities

* chore: Remove empty lines in solanaActions root controller

* feat: Support headers in http response builder fns

This eliminates the need to manually construct responses when we need to
return headers.

* feat: Upload pool images to S3

When user creates a pool with a custom image, we now save the image to
S3 and use that URL to create the pool on Solana.

For non-gif images, we convert it to png type, which ensures compatibility
with dialect, and also eliminates XSS vulnerabilities. GIF's we upload to
S3 as is.

* refactor: Abstract `processAndUploadImage` to a separate function

This keeps the controller logic much leaner

* fix: Remove AWS_REGION from lambda env

This is apparently a reserved env variable, which is injected automatically

* fix: Replace `sharp` to fix lambda runtime issue

Sharp is a native library, so we are facing some mismatch between our
build/dev env (ubuntu) and prod (Amazon Linux 2 i.e. lambda) in terms of
dependencies.

In order to work around this, we now use the file-type package. The
`createPoolTx` handler logic was also modified to not try and convert
images to png, but rather use an allow-list of image types.

We are checking the image buffer to verify that it is indeed the specified
type instead of relying on MIME types like earlier.  Also we sanitize svgs.
These two measures help mitigate XSS attacks from S3 stored images.

* feat: Sanitize SVG images provided to `generateCreatePoolTx``

This was forgotten earlier. Sanitizing the images ensures that no XSS
attack occurs

* feat: URL encode S3 object name

Often times the S3 object name is dynamically set, while the path is
hardcoded. We now encode the S3 object name to ensure reliable URL
creation.
Otherwise S3 usually substitutes certain special characters (like ' '
with '+'), which can potentially break the URL.

* feat: Modify `saveImage` handler to use `S3Service`

The tests were also updated